requirements.txt
streamlit
pandas
numpy
pymongo
werkzeug
langchain
langchain-google-genai
spacy
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
keybert
rake-nltk
sentence-transformers
python-pptx
python-docx
pymupdf
pymongo
dnspython
python-dotenv
